Fix non-ascii tracebacks in FancyReporter.
Tracebacks may contain non-ASCII characters, eg. in literal Unicode
strings. I guess that the encoding of these characters is the same as in
the source file, which is not easy to dertermine and may not even be the
same accross files.

Tell Pygments to guess the encoding, ie. try UTF-8 and fall back on
latin-1. Previously it just assumed latin-1. This only fixes the bug
when all source files are UTF-8, but that is recommended anyway.